Fabric.js 2.4.6 implementation
I used fabric.js paths with are similar to SVG paths. This allowed me to
remove the need for the apache library. I also simplified the
development workflow because I couldn't figure out why all the overhead
was needed or useful... or honestly how to even start with all that. A
simple text editor and broswer worked just fine.

This still needs much work! Only the 4 simple examples shown work, it
does not work if the left-most box is lower on the screen than the right
box, nor does it work for arrows from right to left yet. This also does
not recalculate paths when objects move.